Fix: No cargaba Grh's > 32767

+++ Codigo/clsDX8Engine.bas
@@ -388,7 +388,7 @@ Function InMapBounds(X As Integer, Y As Integer) As Boolean
 
 End Function
 
-Public Sub Draw_GrhIndexMiniMap(ByVal grh_index As Integer, _
+Public Sub Draw_GrhIndexMiniMap(ByVal grh_index As Long, _
                                 ByVal X As Integer, _
                                 ByVal Y As Integer, _
                                 ByVal PixelWidth As Long, _
